+++ Sources/VComponents/Components/Buttons/Rounded Button/VRoundedButtonLabel.swift
@@ -9,7 +9,6 @@ import SwiftUI
 
 // MARK: - V Rounded Button Label
 @available(tvOS, unavailable)
-@available(watchOS, unavailable)
 enum VRoundedButtonLabel<Label> where Label: View {
     case title(title: String)
     case icon(icon: Image)