+++ Sources/VComponents/Components/Modals/Modal/VModal.swift
@@ -166,7 +166,7 @@ struct VModal<Content>: View
     
     private var closeButtonCompensator: some View {
         Spacer()
-            .frame(width: uiModel.layout.closeButtonSubUIModel.dimension)
+            .frame(width: uiModel.layout.closeButtonSubUIModel.size.width)
     }
 
     // MARK: Animations